transferred class prompt_template in different file
